*** Changed variable names for database config
All of the mysql libraries that I can find for NodeJS use; host, user, database and password as variable names for connecting to a database.  This commit brings Trongate into line with what developers from the Node community are putting out.  This frees me up from having to write awkward string replacement code on the desktop app.  This change also opens up the potential for Trongate to be integrated with additional database types in the future.