Errors with more than one Date-Picker / Clear Date
Changed the function "disableDatePickerInputs". 
The Line 670: var targetEl = datePickerInputs[i]; was responsible for the fact that targetEl was always pointing to the last Date-Input in a form. This was the main problem solved by using  "ev.target" instead inside the event handler.

Next problem was changing focus between two Date-Inout-Fields. The "mousedown" event of the next field occurs bevore the  "blure" event of the old field. That leads to storing the wrong value in the variable "originalValue". This was solved bei changing "mousedown" to "focus" event. Inside the blur event I am using now the new parseDate function to test for a valid Date.

The last Change was to have the possibility to clear a date input, for example in case of an optional field. Solved by testing for the  "Backspace" or "Delete" key in combination with testing for "ev.target.value.length == 0". So selecting the whole date-string and press Backspace works perfekt to clear a once selected Date.