Minor fix to CSRF token
It appeared that the newer version of form_helper.php is generating a new CSRF token each time form_close() is called. This was the reason for the issue we were facing with the single image uploader, as a new token was being generated before the form was submitted, leading to a mismatch during validation. Therefore the fix would be to only generate a CSRF token once per session.  Thanks to michidesign for spotting this and letting us know via the Help_Bar on Trongate.io > thread https://trongate.io/help_bar_threads/display/KSsFCTJUTDrkWz56