Added ability to use query parameters to segment 1 & 2 of the URL
This has been a long standing issue where if you tried to use query parameters strings on segment 1 or segment 2 of the URL it would result in the red 404 ERROR : PAGE NOT FOUND being displayed.  e.g. http://localhost/your_trongate_app/?x=1

Recently there was a question on the Help_bar by investobd; where this user had a Trongate site they wanted to share on Facebook.  This resulted in Facebook adding "?fbclid= token" to the URL and breaking the link.

Traditionally Trongate reserved segment(1) for the module name and segment(2) for the method name and then allowed segment(3) and beyond to use 'path parameters' or  you could also use query parameters on segment(3).

This fix:
1. Allows you to use Trongate as before with NO BREAKING CHANGES
2. Allows you now to use query parameter strings on segment 1
3. Allows you now to use query parameter strings on segment 2