Fix type error when index is 0
As reported on the Trongate Help Bar, user mjim found that when changing the 'Records Per Page' drop down to '10' on 'trongate_pages/manage' it would trigger a type error.

This change fixes that problem by removing the type hinting and ensuring $selected_index is always an interger when setting the index to the $_SESSION variable 'selected_per_page'.

Also removed the if block, that was checking for is_numeric, which is not needed and a was setting the wrong index from the default value of 20.