Add explicit return null to query_bind method
Freshly installed app: on the api/explorer/trongate_pages page there is a fatal error: Return value must be of type object|array|null, none returned.

I have added an explicit return of null to the end of the method so that it returns null instead of none. The API Explorer is now back to normal.